Fixes to SPOSiteGroup Extraction

+++ Modules/Office365DSC/DSCResources/MSFT_SPOSiteGroup/MSFT_SPOSiteGroup.psm1
@@ -89,7 +89,7 @@ function Get-TargetResource
     {
         if ($_.Exception -like '*Access denied*')
         {
-            Write-Warning -Message "The specified account does not have access to the permissions list for {$Url}"
+            Write-Warning -Message "The specified account does not have access to the permissions list"
             return $nullReturn
         }
     }
@@ -97,6 +97,10 @@ function Get-TargetResource
     foreach ($entry in $sitePermissions.RoleTypeKind)
     {
         $permissions += $entry.ToString()
+    }
+    if ($permissions.Length -eq 0)
+    {
+        return $nullReturn
     }
     return @{
         Url                = $Url
@@ -345,7 +349,7 @@ function Export-TargetResource
             }
             catch
             {
-                Write-Warning -Message "The specified account does not have access to the permissions list for {$Url}"
+                Write-Warning -Message "The specified account does not have access to the permissions list for {$($siteGroup.Title)}"
                 break
             }
             $params = @{
@@ -356,13 +360,16 @@ function Export-TargetResource
             try
             {
                 $result = Get-TargetResource @params
-                $result = Remove-NullEntriesFromHashtable -Hash $result
-                $result.GlobalAdminAccount = Resolve-Credentials -UserName "globaladmin"
-                $content += "        SPOSiteGroup " + (New-GUID).ToString() + "`r`n"
-                $content += "        {`r`n"
-                $currentDSCBlock = Get-DSCBlock -Params $result -ModulePath $PSScriptRoot
-                $content += Convert-DSCStringParamToVariable -DSCBlock $currentDSCBlock -ParameterName "GlobalAdminAccount"
-                $content += "        }`r`n"
+                if ($result.Ensure -eq 'Present')
+                {
+                    $result = Remove-NullEntriesFromHashtable -Hash $result
+                    $result.GlobalAdminAccount = Resolve-Credentials -UserName "globaladmin"
+                    $content += "        SPOSiteGroup " + (New-GUID).ToString() + "`r`n"
+                    $content += "        {`r`n"
+                    $currentDSCBlock = Get-DSCBlock -Params $result -ModulePath $PSScriptRoot
+                    $content += Convert-DSCStringParamToVariable -DSCBlock $currentDSCBlock -ParameterName "GlobalAdminAccount"
+                    $content += "        }`r`n"
+                }
             }
             catch
             {
